Crate virtual_nic

This crate defines a struct that enables language-level virtualization of the NIC. A VirtualNIC contains a subset of physical NIC resources that are sufficient for an application, that has mutable access to a VirtualNIC, to send and receive packets without kernel mediation. The resources of a VirtualNIC are a set of RxQueues and TxQueues that are passed to it by a physical NIC. On creation of a VirtualNIC, hardware filters are set to route packets sent to a specified IP address to the VirtualNIC queues. When a VirtualNIC is dropped, its resources are returned to the physical NIC.

Structs

VirtualNic

A structure that contains a set of RxQueues and TxQueues that can be used to send and receive packets.