Struct virtual_nic::VirtualNic

pub struct VirtualNic<S: RxQueueRegisters + 'static, T: RxDescriptor + 'static, U: TxQueueRegisters + 'static, V: TxDescriptor + 'static> { /* private fields */ }

A structure that contains a set of RxQueues and TxQueues that can be used to send and receive packets.

Implementations

impl<S: RxQueueRegisters, T: RxDescriptor, U: TxQueueRegisters, V: TxDescriptor> VirtualNic<S, T, U, V>

pub fn new(
rx_queues: Vec<RxQueue<S, T>>,
default_rx_queue: usize,
tx_queues: Vec<TxQueue<U, V>>,
default_tx_queue: usize,
mac_address: [u8; 6],
physical_nic_ref: &'static MutexIrqSafe<dyn PhysicalNic<S, T, U, V>>
) -> Result<VirtualNic<S, T, U, V>, &'static str>

Create a new VirtualNIC with the given parameters. For now we require that there is at least one Rx and one Tx queue.

pub fn id(&self) -> u8

pub fn send_packet_on_queue(
&mut self,
qid: usize,
transmit_buffer: TransmitBuffer
) -> Result<(), &'static str>

Send a packet on the specified queue.